+++ src/main/java/com/clickhouse/kafka/connect/sink/db/ClickHouseWriter.java
@@ -194,7 +194,7 @@ public class ClickHouseWriter implements DBWriter {
         }
     }
 
-    private boolean validateDataSchema(Table table, Record record, boolean onlyFieldsName) {
+    protected boolean validateDataSchema(Table table, Record record, boolean onlyFieldsName) {
         boolean validSchema = true;
         for (Column col : table.getRootColumnsList()) {
             String colName = col.getName();
@@ -246,7 +246,7 @@ public class ClickHouseWriter implements DBWriter {
         return validSchema;
     }
 
-    private void doWriteDates(Type type, ClickHousePipedOutputStream stream, Data value, int precision) throws IOException {
+    protected void doWriteDates(Type type, ClickHousePipedOutputStream stream, Data value, int precision) throws IOException {
         // TODO: develop more specific tests to have better coverage
         if (value.getObject() == null) {
             BinaryStreamUtils.writeNull(stream);
@@ -361,7 +361,7 @@ public class ClickHouseWriter implements DBWriter {
         }
     }
 
-    private void doWriteColValue(Column col, ClickHousePipedOutputStream stream, Data value, boolean defaultsSupport) throws IOException {
+    protected void doWriteColValue(Column col, ClickHousePipedOutputStream stream, Data value, boolean defaultsSupport) throws IOException {
         Type columnType = col.getType();
 
         switch (columnType) {
@@ -502,7 +502,7 @@ public class ClickHouseWriter implements DBWriter {
         }
     }
 
-    private void doWriteFixedString(Type columnType, ClickHousePipedOutputStream stream, Object value, int length) throws IOException {
+    protected void doWriteFixedString(Type columnType, ClickHousePipedOutputStream stream, Object value, int length) throws IOException {
         LOGGER.trace("Writing fixed string type: {}, value: {}", columnType, value);
 
         if (value == null) {
@@ -523,7 +523,7 @@ public class ClickHouseWriter implements DBWriter {
             }
         }
     }
-    private void doWritePrimitive(Type columnType, Schema.Type dataType, ClickHousePipedOutputStream stream, Object value, Column col) throws IOException {
+    protected void doWritePrimitive(Type columnType, Schema.Type dataType, ClickHousePipedOutputStream stream, Object value, Column col) throws IOException {
         LOGGER.trace("Writing primitive type: {}, value: {}", columnType, value);
 
         if (value == null) {
@@ -597,7 +597,7 @@ public class ClickHouseWriter implements DBWriter {
     }
 
 
-    private void doWriteCol(Record record, Column col, ClickHousePipedOutputStream stream, boolean defaultsSupport) throws IOException {
+    protected void doWriteCol(Record record, Column col, ClickHousePipedOutputStream stream, boolean defaultsSupport) throws IOException {
         LOGGER.trace("Writing column {} to stream", col.getName());
         LOGGER.trace("Column type is {}", col.getType());
         String name = col.getName();
@@ -670,7 +670,7 @@ public class ClickHouseWriter implements DBWriter {
         String database = first.getDatabase();
 
         if (!validateDataSchema(table, first, false))
-            throw new RuntimeException();
+            throw new RuntimeException("Data schema validation failed.");
         // Let's test first record
         // Do we have all elements from the table inside the record